+++ oxe-web-admin/src/component/item/Image.jsx
@@ -3,7 +3,7 @@ import "./Image.css";
 import { NotificationManager as nm } from "react-notifications";
 import Popup from "reactjs-popup";
 import { getApiURL } from "../../utils/env.jsx";
-import { validateWord } from "../../utils/re.jsx";
+import { validateWords } from "../../utils/re.jsx";
 import Chip from "../button/Chip.jsx";
 import Message from "../box/Message.jsx";
 import { getRequest, postRequest } from "../../utils/request.jsx";
@@ -109,11 +109,12 @@ export default class Image extends Component {
 	addKeyword(word) {
 		if (this.state.image) {
 			let words;
+			const wordsToAdd = word.split(" ").filter((w) => w.length > 2);
 
 			if (this.state.image.keywords) {
-				words = (this.state.image.keywords + " " + word).toLowerCase();
+				words = (this.state.image.keywords + " " + wordsToAdd.join(" ")).toLowerCase();
 			} else {
-				words = word.toLowerCase();
+				words = wordsToAdd.join(" ").toLowerCase();
 			}
 
 			this.updateImage(this.state.image.id, words);
@@ -233,12 +234,12 @@ export default class Image extends Component {
 								<div className="col-md-12">
 									<input
 										autoFocus
-										className={!validateWord(this.state.word) ? "FormLine-wrong-format" : ""}
+										className={!validateWords(this.state.word) ? "FormLine-wrong-format" : ""}
 										type={"text"}
 										value={this.state.word}
 										onChange={(v) => this.setState({ word: v.target.value })}
 										onKeyPress={(e) => {
-											if (e.key === "Enter" && validateWord(this.state.word)) {
+											if (e.key === "Enter" && validateWords(this.state.word)) {
 												this.addKeyword(this.state.word);
 											}
 										}}
@@ -248,7 +249,7 @@ export default class Image extends Component {
 								<div className="col-md-12 right-buttons">
 									<button
 										onClick={() => this.addKeyword(this.state.word)}
-										disabled={!validateWord(this.state.word)}
+										disabled={!validateWords(this.state.word)}
 									>
 										Add keyword
 									</button>