+++ oxe-web-admin/src/utils/re.jsx
@@ -24,6 +24,11 @@ export function validateWord(value) {
 	return value.match(/^([0-9a-zA-Z-]){3,100}$/);
 }
 
+export function validateWords(value) {
+	if (typeof value === "undefined" || value === null || value.length === 0) return false;
+	return value.match(/^([0-9a-zA-Z-\s]){3,100}$/);
+}
+
 export function validateUrl(value) {
 	if (typeof value === "undefined" || value === null || value.length === 0) return false;
 	return value.match(/^https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)$/);